backend/src/requirements.txt
PyJWT>=2.0
psycopg[binary]>=3.2
SQLAlchemy>=2.0
alembic>=1.13
